requirements.txt
APScheduler==3.10.4
asgiref==3.8.1
colorama==0.4.6
Django==5.0.6
django-appconf==1.0.6
django-cors-headers==4.4.0
django-debug-toolbar==4.4.2
django-select2==8.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==10.3.0
PyJWT==2.8.0
pypng==0.20220715.0
pytz==2024.1
qrcode==7.4.2
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
